+++ src/main.py
@@ -4,7 +4,7 @@ import sys
 
 # import arg_parser
 # import parser.command as cmd
-import parser.option as cmd
+import parse.option as cmd
 
 # def main():
 #     parse_input = arg_parser.ArgParser()
