+++ src/parse/preprocess.py
@@ -14,6 +14,7 @@
 # Do I need a preprocess step?
 # Should I parse at this step and avpid preprocess?
 
+
 class Preprocess:
 
     def __init__(self) -> None:
